refactor: Shared rules should live in a shared baseline across PHP versions. Any overrides can be declared in the specific class.
